Correção Relatorios_Pagamento

+++ backend-web-api/src/main/java/br/com/webdois/backend_web_api/dtos/ReceitaPorCategoria.java
@@ -1,11 +1,10 @@
 package br.com.webdois.backend_web_api.dtos;
 
-import java.math.BigDecimal;
-
 public interface ReceitaPorCategoria {
+
     Long getCategoriaId();
 
     String getCategoriaNome();
 
-    BigDecimal getTotal();
+    Double getTotal();
 }
