+++ backend-web-api/src/main/java/br/com/webdois/backend_web_api/dtos/ReceitaPorDia.java
@@ -1,10 +1,10 @@
 package br.com.webdois.backend_web_api.dtos;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public interface ReceitaPorDia {
+
     LocalDate getDia();
 
-    BigDecimal getTotal();
+    Double getTotal();
 }
